pyproject.toml
[tool.poetry]
name = "pp"
version = "0.1.0"
description = ""
authors = ["Kornilin5 <sasha_kornilin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc9"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
allure-pytest = "^2.13.5"
pytest-repeat = "^0.9.3"
requests = "^2.31.0"
faker = "^24.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
